requirements.txt
flask==0.9
flask-login
flask-mail
sqlalchemy==0.7.9
flask-sqlalchemy==0.16
sqlalchemy-migrate
flask-whooshalchemy==0.54a
flask-wtf==0.8.4
flup
